pyproject.toml
[project]
name = "kunsei"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gpiozero>=2.0.1",
    "langchain>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "lgpio>=0.2.2.0",
    "mcp>=1.22.0",
    "openai>=2.8.1",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
]
